% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{aggregate_clusters}
\alias{aggregate_clusters}
\title{Represent each sample by its cluster composition}
\usage{
aggregate_clusters(kasumi.clusters)
}
\arguments{
\item{kasumi.clusters}{clusters calculated from the extracted representation}
}
\description{
Represent each sample by its cluster composition
}
