% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_kasumi_cluster}
\alias{collect_kasumi_cluster}
\title{Collect Kasumi cluster description}
\usage{
collect_kasumi_cluster(kasumi.clusters, cluster, db.file)
}
\description{
Collect Kasumi cluster description
}
