% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_kwc}
\alias{collect_kwc}
\title{Collect KWC data}
\usage{
collect_kwc(db.file, sample.pattern = ".")
}
\arguments{
\item{db.file}{path to database file with raw results from
\code{\link{run_kasumi}()}.}

\item{sample.pattern}{a regex pattern to match sample names.}
}
\value{
Long format \code{tibble} with the relative abundance of each
\var{Target} in each window and \var{sample}.
}
\description{
Collect KWC composition data per window produced by \code{\link{run_kasumi}()}.
}
