% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downstream_classify}
\alias{downstream_classify}
\title{Use Kasumi representation for downstream classification}
\usage{
downstream_classify(kasumi.clusters, target, dir = "auto", seed = 1)
}
\description{
Use Kasumi representation for downstream classification
}
