% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{extract_clusters}
\alias{extract_clusters}
\title{Extract cluster based representation}
\usage{
extract_clusters(
  kasumi.representation,
  type = c("leiden", "kmeans"),
  sim = 0.8,
  res = 0.8,
  k = 10,
  seed = 1
)
}
\description{
Extract cluster based representation
}
