% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{extract_representation}
\alias{extract_representation}
\title{Extract first kasumi representation}
\usage{
extract_representation(kasumi.results, cutoff = 0, trim = 1)
}
\description{
Importance of relationships or KWC composition
}
