% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{persistent_clusters}
\alias{persistent_clusters}
\title{Identify persistent clusters}
\usage{
persistent_clusters(kasumi.clusters, parameter = 0.1)
}
\description{
Representation is either raw kasumi.clusters or aggregated kasumi.clusters
}
