% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{create_initial_view}
\alias{create_view}
\alias{add_views}
\alias{add_juxtaview}
\alias{add_paraview}
\alias{remove_views}
\alias{filter_views}
\alias{select_markers}
\alias{rename_view}
\alias{collect_results}
\alias{extract_signature}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{mistyR}{\code{\link[mistyR]{add_juxtaview}}, \code{\link[mistyR]{add_paraview}}, \code{\link[mistyR]{add_views}}, \code{\link[mistyR]{collect_results}}, \code{\link[mistyR]{create_initial_view}}, \code{\link[mistyR]{create_view}}, \code{\link[mistyR]{extract_signature}}, \code{\link[mistyR]{filter_views}}, \code{\link[mistyR]{remove_views}}, \code{\link[mistyR]{rename_view}}, \code{\link[mistyR]{select_markers}}}
}}

