% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasumi.R
\name{run_kasumi}
\alias{run_kasumi}
\title{Train Kasumi models}
\usage{
run_kasumi(
  views,
  positions,
  window,
  overlap = 50,
  sample.id = "sample",
  results.db = paste0(sample.id, ".sqm"),
  minu = 50,
  ...
)
}
\arguments{
\item{views}{view composition.}

\item{positions}{a \code{data.frame}, \code{tibble} or a \code{matrix}
with named coordinates in columns and rows for each spatial unit ordered
as in the intraview.}

\item{window}{size of the window.}

\item{overlap}{overlap of consecutive windows (percentage).}

\item{sample.id}{id of the sample.}

\item{results.db}{path to the database file to store the results.}

\item{minu}{minimum number of spatial units in the window.}

\item{...}{all other parameters are passed to \code{\link{run_misty}()}.}
}
\value{
Path to the result folder(s) that can be passed to
\code{\link{collect_results}()}.
}
\description{
Train local multi-view models by sliding a window across the sample as captured by the
view composition.
}
\examples{
# Create a view composition of an intraview and a paraview with radius 10 then
# run Kasumi for a single sample.

library(dplyr)

# get the expression data
data("synthetic", package="mistyR")
expr <- synthetic[[1]] \%>\% select(-c(row, col, type))
# get the coordinates for each cell
pos <- synthetic[[1]] \%>\% select(row, col)

# compose
kasumi.views <- create_initial_view(expr) \%>\% add_paraview(pos, l = 10)

# run with a window of size 100
run_kasumi(kasumi.views, pos, window = 100)

}
